requirements.txt
Chroma
chroma-hnswlib
chromadb
gradio
ipykernel
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-tavily
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langmem
langsmith
numpy
# ==1.26.4
openai
pandas
# ==2.2.2
pydantic
pydantic_core
pyodbc
pypdf
pyprojroot
PyYAML
SQLAlchemy